Widen the two acquisition-path hops dev extracted after the merge
pyright on the merged tree found a float offset flowing into two int-annotated
parameters. Both are real, and both are artefacts of the merge rather than of
this branch: dev extracted _resolve_run_knobs out of simulate_voting_iterations
and threshold_from_fold_orderings out of the conformal rule AFTER this branch had
widened their callers, so the new hops kept the old annotation.

Exactly the kind of thing a conflict-free merge hides and a type gate does not.

+++ vtscore/eval/voting_iterations.py
@@ -1031,7 +1031,7 @@ def _resolve_run_knobs(
     strategy: str,
     skyline_arms: Optional[list[str]],
     emit_calibration_metrics: bool,
-    acq_inclusion_offset: int,
+    acq_inclusion_offset: float,
     acq_rank_percentile: Optional[float],
     head: Optional[str],
     trainer: str,
